Replaced old view with the new one
