+++ webapp/src/App.jsx
@@ -7,13 +7,12 @@ import Signup from "./views/Signup";
 import About from "./views/About";
 import Ranking from "./views/Ranking";
 import Menu from "./views/Menu";
-import GameV2 from "./views/GameV2";
+import Game from "./views/Game";
 import Account from "./views/Account";
 import Error from "./views/Error";
 import { ThemeProvider, createTheme } from "@mui/material/styles";
 import Particles from "./views/components/Particles";
 import React, { useState, useEffect } from "react";
-import Endgame from "./views/Endgame";
 import axios from "axios";
 
 import configDefault from "./views/components/config/particles-config.json";
@@ -123,7 +122,7 @@ export default function App() {
           <Route path="/about" element={<About />} />
           <Route path="/ranking" element={<Ranking />} />
           <Route path="/menu" element={<Menu />} />
-          <Route path="/game/:category" element={<GameV2 />} />
+          <Route path="/game/:category" element={<Game />} />
           <Route path="/account" element={<Account />} />
           <Route path="*" element={<Error />} />
         </Routes>
